+++ src/client.ts
@@ -19,6 +19,7 @@ import { Institutional } from "./resources/institutional.js";
 import { KB } from "./resources/kb.js";
 import { MarketMoodResource } from "./resources/marketMood.js";
 import { MarketSummaryResource } from "./resources/marketSummary.js";
+import { Screener } from "./resources/screener.js";
 import { Stocks } from "./resources/stocks.js";
 import { Indexes } from "./resources/indexes.js";
 import { Trackers } from "./resources/trackers.js";
@@ -101,6 +102,7 @@ export class SentiSense implements APIClient {
   readonly trackers: Trackers;
   readonly calendar: Calendar;
   readonly earnings: Earnings;
+  readonly screener: Screener;
 
   constructor(options: SentiSenseOptions = {}) {
     this.apiKey = options.apiKey;
@@ -124,6 +126,7 @@ export class SentiSense implements APIClient {
     this.trackers = new Trackers(this);
     this.calendar = new Calendar(this);
     this.earnings = new Earnings(this);
+    this.screener = new Screener(this);
   }
 
   /** @internal */